422: update solution

+++ algorithm_0000-0999/422-Valid Word Square/main.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
     bool validWordSquare(vector<string>& words) {
-        for (size_t r = 0; r < words.size(); ++r) {
-            if (words.size() <words[r].size()) {
+        for (int r = 0; r < words.size(); ++r) {
+            if (words.size() < words[r].size()) {
                 return false;
             }
-            for (size_t c = 0; c < words[r].size(); ++c) {
+            for (int c = 0; c < words[r].size(); ++c) {
                 if (words[r][c] != words[c][r]) {
                     return false;
                 }
